Update README.md
The Corner Grocer Item Frequency Tracker is a C++ console-based application developed to help a local grocery store track and analyze how frequently items are purchased. This program reads a list of grocery items from an input file, tallies the number of times each item appears, and offers the user a simple menu interface to interact with this data.